requirements.txt
discord.py==2.3.2
pyyaml==6.0.1
twscrape==0.11.1